Add reset and recover password logics

+++ backend/app/api/routes/users.py
@@ -1,11 +1,12 @@
 # Routes for user operations
-from typing import Annotated
+import uuid
+from typing import Annotated, Any
 
-from fastapi import APIRouter, HTTPException, Query, status
+from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlmodel import select
 
-from app.core.deps import SessionDep
-from app.models.schemas import User, UserCreate, UserPublic, UserUpdate
+from app.core.deps import CurrentUser, SessionDep, get_current_active_superuser
+from app.models.schemas import Item, Message, User, UserCreate, UserPublic, UserUpdate
 
 router = APIRouter(prefix="", tags=["users"])
 
@@ -25,11 +26,24 @@ def create_user(user: UserCreate, session: SessionDep):
 @router.get("/users/", response_model=list[UserPublic])
 def read_users(
     session: SessionDep,
-    offset: int = 0,
+    skip: int = 0,
     limit: Annotated[int, Query(le=100)] = 100,
-):
-    users = session.exec(select(User).offset(offset).limit(limit)).all()
-    return users
+) -> Any:
+    """Retrives users.
+
+    Args:
+        session (SessionDep): _description_
+        offset (int, optional): _description_. Defaults to 0.
+        limit (Annotated[int, Query, optional): _description_. Defaults to 100)]=100.
+
+    Returns:
+        Any: _description_
+    """
+    count_statment = select(func.count()).select_from(User)
+    count = session.exec(count_statment).one()
+
+    users = session.exec(select(User).offset(skip).limit(limit)).all()
+    return UserPublic(data=users, count=count)
 
 
 # Read user by Id
@@ -61,13 +75,35 @@ def update_user(user_id: int, user: UserUpdate, session: SessionDep):
 
 
 # Delete user
-@router.delete("/users/{user_id}")
-def delete_user(user_id: int, session: SessionDep):
+@router.delete("/users/{user_id}", dependencies=[Depends(get_current_active_superuser)])
+def delete_user(
+    user_id: uuid.UUID, current_user: CurrentUser, session: SessionDep
+) -> Message:
+    """_summary_
+
+    Args:
+        user_id (uuid.UUID): _description_
+        current_user (CurrentUser): _description_
+        session (SessionDep): _description_
+
+    Raises:
+        HTTPException: user not found
+
+    Returns:
+        Message: User deleted successfully
+    """
     user = session.get(User, user_id)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="user not found"
         )
+    if user == current_user:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Super users are not allowed to delete themselves",
+        )
+    statement = delete(Item).where(col(Item.owner_id) == user_id)
+    session.exec(statement)
     session.delete(user)
     session.commit()
-    return {"ok": True}
+    return Message(message="User deleted successfully")
